requirements.txt
absl-py==2.1.0
acres==0.5.0
aggdraw==1.3.19
alembic==1.14.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
build==1.3.0
cachetools==6.2.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
chromadb==1.3.5
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.6
configobj==5.0.9
configparser==7.2.0
contourpy==1.2.1
cycler==0.12.1
Cython==3.1.6
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
durationpy==0.10
etelemetry==0.3.1
executing==2.0.1
fake-useragent==2.2.0
fastapi==0.121.0
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
flask-cors==6.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
fsspec==2025.10.0
gast==0.6.0
gdown==5.2.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.1.1
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.15.1
httpcore==1.0.5
httplib2==0.31.0
httptools==0.7.1
httpx==0.27.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.8
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.11.3
kiwisolver==1.4.5
kubernetes==34.1.0
lab==8.3
langchain==1.1.0
langchain-core==1.1.0
langchain-google-genai==3.2.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.47
libclang==18.1.1
looseversion==1.3.0
lxml==6.0.2
Mako==1.3.8
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.0.2
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
multipart==1.3.0
namex==0.1.0
narwhals==1.30.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opt-einsum==3.3.0
optree==0.17.0
orjson==3.11.4
ormsgpack==1.12.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
patsy==1.0.2
pillow==10.4.0
platformdirs==4.2.2
plotly==6.0.0
pmdarima==2.0.4
posthog==5.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==6.33.0
prov==2.1.1
psutil==6.0.0
pure_eval==0.2.3
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycaw==20240210
pycparser==2.22
pydantic==2.12.4
pydantic_core==2.41.5
pydot==4.0.1
pygame==2.6.0
Pygments==2.18.0
pyjokes==0.7.2
PyMuPDF==1.26.6
pyparsing==3.1.2
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==2.0.7
pytz==2024.2
pywin32==306
pywinpty==2.0.13
pyxnat==1.6.4
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==7.4.2
qtconsole==5.5.2
QtPy==2.4.1
rdflib==7.4.0
referencing==0.35.1
regex==2025.11.3
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.2.0
rpds-py==0.20.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.14.0
seaborn==0.13.2
selenium==4.28.1
Send2Trash==1.8.3
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.49.3
statsmodels==0.14.5
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado==0.18.1
tf_keras==2.20.1
threadpoolctl==3.6.0
tinycss2==1.3.0
tokenizers==0.22.1
torch==2.9.0+cpu
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.57.1
trio==0.28.0
trio-websocket==0.11.1
txt2tags==3.9
typer==0.20.0
types-python-dateutil==2.9.0.20240821
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
undetected-chromedriver==3.5.5
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.38.0
vboxapi==1.0
virtualenv==20.26.6
visualkeras==0.2.0
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.8.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==2.0.0
wsproto==1.2.0
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0